app/requirements.txt
flask==2.2.2
flask_debugtoolbar==0.13.1
flask-compress==1.13
flask_login==0.6.2
Flask-Migrate==4.0.1
flask-paginate==2022.1.8
flask-simple-captcha==5.5.5
flask_sqlalchemy==3.0.2
Flask-WTF==1.1.1
gevent==22.10.2
google-api-python-client==2.80.0
google_auth_oauthlib==1.0.0
gunicorn==20.1.0
oauth2client==4.1.3
openpyxl==3.0.10
pandas==1.5.2
PyPDF2==3.0.1
psycopg2-binary==2.9.5
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==0.21.0
pyTelegramBotAPI==4.10.0
rarfile==4.1
reportlab==4.0.9
requests==2.28.1
redis==4.5.5
Werkzeug==2.2.2
xlsxwriter==3.0.3
xlrd==2.0.1
sib-api-v3-sdk==7.5.0
rq-scheduler==0.13.1
loguru==0.7.2
elastic-apm[flask]==6.20.0
xtea==0.7.1
